editor: use one InteractiveCanvasViewer across all pages
